stat_mean_ci and stat_median_ci to handle edge case n = 1. (#1074)
* stat_mean_ci and stat_median_ci to handle edge case n = 1.

I updated two functions: stat_mean_ci and stat_median_ci to return correct types of objects for ymin and ymax columns for input series of length 1 (after NAs removal). Namely, in such a case, both functions should return NA_real_ instead of NA (logical) or Inf. 

Furthermore, I updated the code for stat_mean_ci to correctly handle extreme combinations of na.rm and n_lim parameters. In particular, I coded the n_lim parameter so that it now correctly denotes the minimum number of observations required to compute a confidence interval for the mean. I think this is the correct meaning of this parameter and this is how it should be coded. NOTE, that with the previous implementation this parameter was understood as follows: the minimum number of observations to compute a confidence interval for the man is equal to n_lim + 1, which I think was just a coding mistake.

I also significantly rearranged the code for stat_median_ci to make it more elegant and clear. I updated conf_level attribute returned by stat_median_ci. It now returns NA_real_ instead of 1 for the cases where confidence intervals are not computed.

I have not adjusted these 2 functions to return NA_real_ for y column for esoteric cases such where input series is just a sequence of NAs. In this case stats::median returns NA (logical).  Similar, stats::mean(NA, na.rm = TRUE) which returns NaN.
I have not handle these esoteric cases as it does not impact clients of our functions, and I wanted to keep outputs of stats::mean and stat::median transparent.

Finally, I created a bunch of test for these functions.

* modification of 2 unit tests for s_summary

Modification of two unit tests for s_summary to reflect changes in stat_mean_ci and stat_median_fun.


In stat_mean_ci - changed back the coding of n_lim parameter. It now has a meaning as before, i.e.: minimum number of required observations to compute confidence bounds is equal to n_lim + 1.

* Update unit tests to reflect changes in stat_median_ci

Update unit tests to reflect changes in stat_median_ci: NA_real_ is returned, instead of +/-Inf.

* Implementing PR approver's suggestions

in stat_mean_ci: 
(1) change the name of the n_lim parameter to n_min as this seems to be more informative
(2) change the meaning of the n_min. It now denotes the minimum no of elements in x required to compute a confidence interval.
(3) minor adjustment to where na.omit is placed.

adjusted unit tests to reflect above changes

* Roxygen update for stat_mean_ci 

Paramter name change from n_lim to n_min in stat_mean_ci function

* change unit test after stat_mean_ci update

change unit test after stat_mean_ci update, after which the (default) minimum number of records required to compute CI is equal to 2 (it was 3 before).

* NEWS.md update

NEWS.md update

* Shorten news entry and put under v0.7.1 heading

Co-authored-by: Jana Stoilova <jana.stoilova@roche.com>