update h_adsl_adlb_merge_using_worst_flag to have all variables (#1089)
* update h_adsl_adlb_merge_using_worst_flag to have all variables

This is to include all variables from adlb to the result returned by h_adsl_adlb_merge_using_worst_flat

* select more variables and update unit tests

This commit keep adsl variables and AVISIT, AVISITN, PARAMCD, PARAM from adlb. For other variables, if adlb records are missing, these variables are missing.

* Update documentation

* update test to make lines shorter

This is to correct the issues in automation tests that lines exceeding 120 characters.

* update unit test to make lines shorter

This is to align each lines in unit test to make them shorter than 120 characters

* update helper function codes and doc

This commit is to 1) modify codes to comply with pipe_continuation_linter in lintr requirement 2) update doc to comply with 120 characters per line requirement

* Fix indent that resulted in automation failure

* adding adsl to the word dictionary

This is to add the adsl word into the dictionary for automation check to pass

* Update R/abnormal_by_worst_grade_by_baseline.R

Co-Authored-By: Stoilova, Jana {MDBB~Mississauga} <jana.stoilova@roche.com>

* Update R/abnormal_by_worst_grade_by_baseline.R

update doc to explain when AVISIT is introduced

Co-Authored-By: Stoilova, Jana {MDBB~Mississauga} <jana.stoilova@roche.com>

* update adlb_f to adlb when constructing grid

1) modify doc to 120 characters each line
2) update adlb_f to adlb when constructing grid with AVISIT

* Update R/abnormal_by_worst_grade_by_baseline.R doc

update doc of helper function

Co-Authored-By: Stoilova, Jana {MDBB~Mississauga} <jana.stoilova@roche.com>

* update worst_flag in example

the by visit example has wrong flag. Set to WGRLOVFL

* update man doc

* remove baseline and screening from AVISIT to build grid

when building the grid, don't want "BASELINE" and "SCREENING" in the grid

* add new argument no_fillin_visits

1) Add new argument no_fillin_visits default to c("SCREENING", "BASELINE")
2) update NEWS.md

* Remove an extra line in NEWS.md

* Update doc param type

following Dev Guide, character is character vector

Co-Authored-By: Stoilova, Jana {MDBB~Mississauga} <jana.stoilova@roche.com>

* Update doc to reflect param type change

Per Dev Guide, no_fillin_visits should be character instead of vector

* skip lintr check for long function name

* dummy commit to restart automation

Co-authored-by: Rosemary Li <liy396@gene.com>
Co-authored-by: Stoilova, Jana {MDBB~Mississauga} <jana.stoilova@roche.com>