fix grob wrapping final (#1305)
Fixes #1287 

I hope this is a wrap (:D). @BFalquet could take a look if it works on
your side? Result may be numerically different but this is the real
solution to the issue.

On another note, we need to consider uniforming `\\n` to `\n` if it
happens. I now know it is not always assumed the same and we have mixed
behavior around the packages. I will check this for {rtables}. I was
able to break wrapping by doing this.